+++ lib/confetti.dart
@@ -3,3 +3,4 @@ library confetti;
 export 'src/confetti.dart';
 export 'src/enums/blast_directionality.dart';
 export 'src/enums/confetti_controller_state.dart';
+export 'src/particle_stats.dart';